+++ plugin-dev/Source/Sentry/Private/SentrySubsystem.cpp
@@ -10,6 +10,7 @@
 #include "SentryUser.h"
 #include "SentryUserFeedback.h"
 #include "SentryBeforeSendHandler.h"
+#include "SentryBeforeBreadcrumbHandler.h"
 #include "SentryTraceSampler.h"
 #include "SentryTransaction.h"
 #include "SentryTransactionContext.h"
@@ -100,6 +101,10 @@ void USentrySubsystem::Initialize()
 	BeforeSendHandler = NewObject<USentryBeforeSendHandler>(this, BeforeSendHandlerClass);
 	check(BeforeSendHandler);
 
+	BeforeBreadcrumbHandler = Settings->BeforeBreadcrumbHandler != nullptr
+		? NewObject<USentryBeforeBreadcrumbHandler>(this, static_cast<UClass*>(Settings->BeforeBreadcrumbHandler))
+		: nullptr;
+
 	const UClass* TraceSamplerClass = Settings->TracesSampler != nullptr
 		? static_cast<UClass*>(Settings->TracesSampler)
 		: USentryTraceSampler::StaticClass();
@@ -107,7 +112,7 @@ void USentrySubsystem::Initialize()
 	TraceSampler = NewObject<USentryTraceSampler>(this, TraceSamplerClass);
 	check(TraceSampler);
 
-	SubsystemNativeImpl->InitWithSettings(Settings, BeforeSendHandler, TraceSampler);
+	SubsystemNativeImpl->InitWithSettings(Settings, BeforeSendHandler, BeforeBreadcrumbHandler, TraceSampler);
 
 	if (!SubsystemNativeImpl->IsEnabled())
 	{