+++ plugin-dev/Source/Sentry/Public/SentrySettings.h
@@ -8,6 +8,7 @@
 #include "SentrySettings.generated.h"
 
 class USentryBeforeSendHandler;
+class USentryBeforeBreadcrumbHandler;
 class USentryTraceSampler;
 
 UENUM(BlueprintType)
@@ -263,6 +264,10 @@ class SENTRY_API USentrySettings : public UObject
 	UPROPERTY(Config, EditAnywhere, BlueprintReadWrite, Category = "General|Hooks",
 		Meta = (DisplayName = "Custom `beforeSend` event handler", ToolTip = "Custom handler for processing events before sending them to Sentry."))
 	TSubclassOf<USentryBeforeSendHandler> BeforeSendHandler;
+
+	UPROPERTY(Config, EditAnywhere, BlueprintReadWrite, Category = "General|Hooks",
+		Meta = (DisplayName = "Custom `beforeBreadcrumb` event handler", ToolTip = "Custom handler for processing breadcrumbs before adding them to the scope."))
+	TSubclassOf<USentryBeforeBreadcrumbHandler> BeforeBreadcrumbHandler;
 
 	UPROPERTY(Config, EditAnywhere, Category = "General|Windows",
 		Meta = (DisplayName = "Override Windows default crash capturing mechanism (UE 5.2+)", ToolTip = "Flag indicating whether to capture crashes automatically on Windows as an alternative to Crash Reporter."))